P4/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
alignment==1.0.10
altgraph==0.16.1
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.8.2
arrayfire==3.6.20181017
arrow==0.13.1
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.0
atomicwrites==1.3.0
attrs==19.1.0
auto-py-to-exe==2.6.1
awscli==1.16.169
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
biopython==1.73
bleach==3.1.0
boto==2.49.0
boto3==1.9.164
botocore==1.12.164
bottle==0.12.16
bottle-websocket==0.2.9
Brlapi==0.6.6
bz2file==0.98
CacheControl==0.12.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.5.2
collections-extended==1.0.2
colorama==0.3.9
command-not-found==0.3
community==1.0.0b1
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.10
datalad==0.11.5
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.5
distro-info==0.18
Django==2.1.4
django-materialize==1.0.1a2
docutils==0.14
docx2txt==0.6
EbookLib==0.15
edlib==1.2.4
Eel==0.10.4
entrypoints==0.3
fake-useragent==0.1.11
fasteners==0.15
Flask==1.0.2
funcy==1.12
future==0.17.1
gast==0.2.2
gensim==3.7.3
gevent==1.4.0
gevent-websocket==0.10.1
gitdb2==2.0.5
GitPython==2.1.11
google==2.0.2
google-pasta==0.1.7
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
hdmedians==0.13
html5lib==0.9999999
httplib2==0.9.2
humanize==0.5.1
idna==2.8
image-slicer==0.2.0
imageai==2.0.2
imageio==2.5.0
imagesize==1.1.0
imgaug==0.2.9
importlib-metadata==0.17
imutils==0.5.2
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isort==4.3.4
itsdangerous==1.1.0
jedi==0.14.0
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsmin==2.2.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.0.2
keyrings.alt==3.1.1
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lda==1.1.0
lockfile==0.12.2
logilab-common==1.4.1
louis==3.5.0
lxml==4.3.1
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mixem==0.1.4
mlxtend==0.15.0.0
mock==3.0.5
monotonic==1.5
more-itertools==7.0.0
msgpack==0.6.1
mxnet==1.3.1
natsort==6.0.0
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.3
nltk==3.3
nose==1.3.7
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.7.0
nwalign==0.3.1
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
opencv-python==4.1.0.25
openshot-qt==2.4.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
patool==1.12
patsy==0.5.1
pbr==4.1.0
pefile==2019.4.18
pexpect==4.7.0
pickle-mixin==1.0.2
pickle-warehouse==0.1.1
pickleshare==0.7.5
Pillow==6.0.0
plotly==2.4.1
pluggy==0.12.0
pocketsphinx==0.1.3
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.4.2
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.8.0
py2exe==0.9.2.2
py4j==0.10.8.1
pyasn1==0.4.5
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyflakes==1.6.0
PyGithub==1.43.7
Pygments==2.4.2
pygobject==3.26.1
PyInstaller==3.4
PyJWT==1.7.1
pyLDAvis==2.1.2
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyopa==0.8.0
pyoperators==0.13.16
pyparsing==2.4.0
PyPDF2==1.26.0
pyRFC3339==1.0
pyrsistent==0.15.2
pyspark==2.4.0
pytesseract==0.2.6
pytest==4.6.2
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
python-Levenshtein==0.12.0
python-pptx==0.6.5
python3-openid==3.1.0
pytz==2019.1
PyWavelets==1.0.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.1
QtAwesome==0.4.4
qtconsole==4.5.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
retrying==1.3.3
roman==2.0.0
rope==0.10.5
rsa==3.4.2
s3transfer==0.2.1
scikit-bio==0.5.5
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.0
scour==0.36
seaborn==0.9.0
SecretStorage==3.1.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
smmap2==2.0.5
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==2.0.0
soupsieve==1.9.1
SpeechRecognition==3.6.3
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.2.6
ssh-import-id==5.7
ssw==0.3.1
statsmodels==0.9.0
stevedore==1.28.0
system-service==0.3
systemd-python==234
tables==3.4.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textract==1.6.1
Theano==1.0.2
times==0.7
torch==0.4.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whichcraft==0.5.2
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xkit==0.0.0
xlrd==1.0.0
XlsxWriter==1.0.5
xlwt==1.3.0
youtube-dl==2018.3.14
zipp==0.5.1
zope.interface==4.3.2
